+++ packages/@leight/file-client/src/sdk/ClientStore.ts
@@ -2,16 +2,19 @@
  Source code containing improved Zustand store stuff for Source support (client-side).
  */
 import {
+    FileFilterSchema,
     FileSchema,
     FileSortSchema,
+    type IFileFilterSchema,
     type IFileSortSchema,
     type IFileSourceSchema
-}                          from "@leight/file";
-import {createSortContext} from "@leight/sort-client";
+}                            from "@leight/file";
+import {createFilterContext} from "@leight/filter-client";
+import {createSortContext}   from "@leight/sort-client";
 import {
     createSourceContext,
     type ISourceProps
-}                          from "@leight/source-client";
+}                            from "@leight/source-client";
 
 export type IFileSource = ISourceProps<IFileSourceSchema>;
 
@@ -21,6 +24,13 @@ export type IFileSource = ISourceProps<IFileSourceSchema>;
 export const FileSourceStore                  = createSourceContext<IFileSourceSchema>({
     name:   "File",
     schema: FileSchema,
+});
+/**
+ * Defines Store for File filtering entities.
+ */
+export const FileFilterStore                  = createFilterContext<IFileFilterSchema>({
+    name:   "FileFilter",
+    schema: FileFilterSchema,
 });
 /**
  * Defines Store for File sorting data.
@@ -33,4 +43,4 @@ export const FileSortStore                    = createSortContext<IFileSortSchem
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_vq0wyc7zcqckmurds0btidkp = true;
+export const $leight_j3dldycyvqlze11odkddg5tm = true;