+++ packages/@leight/file-server/src/sdk/ServerSource.ts
@@ -13,4 +13,4 @@ export class FileSource extends FileBasePrismaSource implements IFileSource {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_jtzbgq6s41bpum3o349w29ts = true;
+export const $leight_qpvbokjewcdz766t02r5wg7r = true;