+++ packages/@leight/file/src/sdk/PrismaSchema.ts
@@ -71,4 +71,4 @@ export const FileQuerySchema                  = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_e7dqia3eo0riz1bzfieykqwh = true;
+export const $leight_wa5rnq9vzacpncbxdjczoh73 = true;