+++ packages/@leight/filter/src/api/index.ts
@@ -1,4 +1,6 @@
-import {z} from "@leight/zod";
+import {type IUseState}   from "@leight/context";
+import {z}                from "@leight/zod";
+import {type IStoreProps} from "@leight/zustand";
 
 export const FilterSchema = z.object({
     id:       z.string().optional(),
@@ -8,3 +10,12 @@ export type IFilterSchema =
     typeof FilterSchema
     | z.ZodType;
 export type IFilter = z.infer<IFilterSchema>;
+
+export type IFilterStoreProps<TFilterSchema extends IFilterSchema> = IStoreProps<{
+    readonly schema: TFilterSchema;
+    readonly filter: z.infer<TFilterSchema>;
+
+    setFilter(filter?: z.infer<TFilterSchema>): void;
+}>
+
+export type IUseFilterState<TFilterSchema extends IFilterSchema> = IUseState<IFilterStoreProps<TFilterSchema>>;