+++ packages/@leight/job-server/src/sdk/ServerPrismaSource.ts
@@ -72,4 +72,4 @@ export class JobBasePrismaSource extends AbstractSource<IJobSourceSchema> {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_zi3mmnpw1t9w1tz6a691c4l0 = true;
+export const $leight_w4z8owaw5iklje7pkcxrhmhd = true;