+++ packages/@leight/job-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const JobSourceProcedure               = withSourceProcedure<IJobSourceSc
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_ey9r14hshl6x86xjpwy7lzn6 = true;
+export const $leight_lzyosei1f6bptga06hfy17mk = true;