+++ packages/@leight/job/src/sdk/PrismaSchema.ts
@@ -72,4 +72,4 @@ export const JobQuerySchema                   = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_ptp2ozbd4q2dh5gc93051dwr = true;
+export const $leight_z78b41c6k597q40wfjzhwjke = true;