+++ packages/@leight/job/src/sdk/SourceSchema.ts
@@ -34,4 +34,4 @@ export interface IJobSourceSchema extends ISourceSchema<
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_wcv2wsf8meiw1op2yz149v8f = true;
+export const $leight_z0et5bol6sxwoidg6pancynt = true;