+++ packages/@leight/user/src/sdk/SourceSchema.ts
@@ -34,4 +34,4 @@ export interface IUserSourceSchema extends ISourceSchema<
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_egapsyui0m7bh0nk11p8heq6 = true;
+export const $leight_mt8avhtxpaz1i5ybsek7l8lt = true;