+++ packages/@leight/context-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/context-client
 
+## 0.5.68
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/context@0.5.23
+    - @leight/utils@0.5.63
+    - @leight/zustand@0.5.60
+
 ## 0.5.67
 
 ### Patch Changes