+++ packages/@leight/file-server/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @leight/file-server
 
+## 0.5.68
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/prisma@0.5.76
+    - @leight/file@0.5.67
+    - @leight/container@0.5.61
+    - @leight/next.js@0.5.26
+    - @leight/next.js-server@0.5.66
+    - @leight/query@0.5.60
+    - @leight/source@0.5.60
+    - @leight/source-server@0.5.60
+    - @leight/trpc-source-server@0.5.14
+    - @leight/utils@0.5.63
+
 ## 0.5.67
 
 ### Patch Changes