+++ packages/@leight/i18n-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/i18n-client
 
+## 0.5.66
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/context-client@0.5.68
+    - @leight/i18n@0.5.66
+    - @leight/utils@0.5.63
+
 ## 0.5.65
 
 ### Patch Changes