+++ packages/@leight/mantine/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leight/mantine
 
+## 0.5.67
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/file-client@0.5.68
+    - @leight/file@0.5.67
+    - @leight/i18n@0.5.66
+    - @leight/i18n-client@0.5.66
+    - @leight/layout@0.5.60
+    - @leight/utils@0.5.63
+    - @leight/utils-client@0.5.64
+
 ## 0.5.66
 
 ### Patch Changes