+++ packages/@leight/source/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @leight/source
 
+## 0.5.60
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/filter@0.5.56
+    - @leight/context@0.5.23
+    - @leight/cursor@0.5.54
+    - @leight/query@0.5.60
+    - @leight/react-query@0.5.23
+    - @leight/sort@0.5.57
+    - @leight/utils@0.5.63
+    - @leight/zod@0.5.6
+    - @leight/zustand@0.5.60
+
 ## 0.5.59
 
 ### Patch Changes