+++ packages/@leight/user/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @leight/user
 
+## 0.5.70
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/filter@0.5.56
+    - @leight/prisma@0.5.76
+    - @leight/container@0.5.61
+    - @leight/query@0.5.60
+    - @leight/react-query@0.5.23
+    - @leight/sort@0.5.57
+    - @leight/source@0.5.60
+    - @leight/zod@0.5.6
+
 ## 0.5.69
 
 ### Patch Changes