+++ packages/@leight/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/utils
 
+## 0.5.63
+
+### Patch Changes
+
+- VErsion bump
+
 ## 0.5.62
 
 ### Patch Changes