+++ packages/@leight/winston/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/winston
 
+## 0.5.59
+
+### Patch Changes
+
+- VErsion bump
+
 ## 0.5.58
 
 ### Patch Changes