+++ packages/@leight/zustand/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/zustand
 
+## 0.5.60
+
+### Patch Changes
+
+- VErsion bump
+- Updated dependencies
+    - @leight/utils@0.5.63
+
 ## 0.5.59
 
 ### Patch Changes